publish firestore-compat
