+++ packages/firestore/package.json
@@ -104,7 +104,7 @@
   "bugs": {
     "url": "https://github.com/firebase/firebase-js-sdk/issues"
   },
-  "typings": "dist/index.d.ts",
+  "typings": "dist/firestore/index.d.ts",
   "nyc": {
     "extension": [
       ".ts"
