+++ scripts/exp/prepare-firestore-for-exp-release.ts
@@ -15,11 +15,20 @@
  * limitations under the License.
  */
 
+import { resolve } from 'path';
 import { projectRoot, readPackageJson } from '../utils';
-import { writeFile as _writeFile, readFile as _readFile } from 'fs';
+import {
+  writeFileSync,
+  statSync,
+  readFileSync,
+  existsSync,
+  rmdirSync,
+  readdirSync,
+  mkdirSync,
+  copyFileSync
+} from 'fs';
 import { promisify } from 'util';
 
-const writeFile = promisify(_writeFile);
 const packagePath = `${projectRoot}/packages/firestore`;
 
 /**
@@ -65,9 +74,108 @@ export async function prepare() {
   ];
 
   // update package.json files
-  await writeFile(
+  writeFileSync(
     `${packagePath}/package.json`,
     `${JSON.stringify(packageJson, null, 2)}\n`,
     { encoding: 'utf-8' }
   );
+
+  // create firestore-compat package for publishing
+  createFirestoreCompatProject();
+}
+
+const FIRESTORE_SRC = resolve(projectRoot, 'package/firestore');
+const FIRESTORE_COMPAT_SRC = resolve(projectRoot, 'packages/firestore/compat');
+const FIRESTORE_COMPAT_DEST = resolve(
+  projectRoot,
+  'packages-exp/firestore-compat'
+);
+const FIRESTORE_COMPAT_BINARY_SRC = resolve(
+  FIRESTORE_COMPAT_SRC,
+  'dist/compat'
+);
+const FIRESTORE_COMPAT_BINARY_DEST = resolve(
+  FIRESTORE_COMPAT_DEST,
+  'packages-exp/firestore-compat/dist'
+);
+
+async function createFirestoreCompatProject() {
+  // remove the dir if it exists
+  if (existsSync(FIRESTORE_COMPAT_DEST)) {
+    rmdirSync(FIRESTORE_COMPAT_DEST, { recursive: true });
+  }
+
+  copyRescursiveSync(FIRESTORE_COMPAT_SRC, FIRESTORE_COMPAT_DEST);
+  copyRescursiveSync(FIRESTORE_COMPAT_BINARY_SRC, FIRESTORE_COMPAT_BINARY_DEST);
+
+  // update root package.json
+  transformFile(
+    resolve(FIRESTORE_COMPAT_DEST, 'package.json'),
+    async content => {
+      const updatedContent = content.replace('../dist/compat', './dist');
+      const compatPkgJson = JSON.parse(updatedContent);
+
+      const firestorePkgJson = await readPackageJson(FIRESTORE_SRC);
+
+      compatPkgJson.dependencies = {
+        ...firestorePkgJson.dependencies,
+        '@firebase/firestore': '0.0.900'
+      };
+
+      compatPkgJson.peerDependencies = {
+        '@firebase/app': '0.x'
+      };
+
+      compatPkgJson.files = ['dist', 'memory', 'bundle'];
+
+      return `${JSON.stringify(compatPkgJson, null, 2)}\n`;
+    }
+  );
+
+  // update bundle/package.json
+  await transformFile(
+    resolve(FIRESTORE_COMPAT_DEST, 'bundle/package.json'),
+    async content => content.replace('../../dist/compat', '../dist')
+  );
+
+  // update memory/package.json
+  await transformFile(
+    resolve(FIRESTORE_COMPAT_DEST, 'memory/package.json'),
+    async content => content.replace('../../dist/compat', '../dist')
+  );
+
+  // update memory/bundle/packages.json
+  await transformFile(
+    resolve(FIRESTORE_COMPAT_DEST, 'memory/bundle/package.json'),
+    async content => content.replace('../../../dist/compat', '../../dist')
+  );
+}
+
+function copyRescursiveSync(src: string, dest: string) {
+  if (!existsSync(src)) {
+    return;
+  }
+
+  const srcStat = statSync(src);
+  const isDirectory = srcStat.isDirectory();
+  if (isDirectory) {
+    mkdirSync(dest);
+    for (const item of readdirSync(src)) {
+      copyRescursiveSync(resolve(src, item), resolve(dest, item));
+    }
+  } else {
+    // do not copy source file
+    if (!src.includes('.ts')) {
+      copyFileSync(src, dest);
+    }
+  }
+}
+
+function transformFile(
+  path: string,
+  transformer: (content: string) => Promise<string>
+) {
+  let content = readFileSync(path, 'utf8');
+
+  writeFileSync(path, transformer(content), { encoding: 'utf8' });
 }
